adding adjustments to CMakelists.txt for proper compilation
